Revert #8456 Move WakeIOThread() functionality to WatchableEventManager
#### Problem

TestPlatformMgr reports errors,
```
CHIP:CSL: System wake event notify failed: OS Error 0x02000009: Bad file descriptor
```
(although it still passes, which obscured the problem).

This might be behind intermittent core dumps in CI.

#### Change overview

Revert #8456 Move WakeIOThread() functionality to WatchableEventManager

#### Testing

No error reported by TestPlatformMgr after reverting.
