+++ src/platform/mbed/PlatformManagerImpl.cpp
@@ -46,9 +46,8 @@ CHIP_ERROR PlatformManagerImpl::_InitChipStack(void)
         mQueue.~EventQueue();
         new (&mQueue) events::EventQueue(event_size * CHIP_DEVICE_CONFIG_MAX_EVENT_QUEUE_SIZE);
 
-        mQueue.background([&](int t) {
-            MbedEventTimeout::AttachTimeout([&] { SystemLayer.WatchableEvents().Signal(); }, std::chrono::milliseconds{ t });
-        });
+        mQueue.background(
+            [&](int t) { MbedEventTimeout::AttachTimeout([&] { SystemLayer.WakeIOThread(); }, std::chrono::milliseconds{ t }); });
 
         // Reinitialize the Mutexes
         mThisStateMutex.~Mutex();
@@ -212,7 +211,7 @@ CHIP_ERROR PlatformManagerImpl::_StopEventLoopTask()
 
     // Wake from select so it unblocks processing
     LockChipStack();
-    SystemLayer.WatchableEvents().Signal();
+    SystemLayer.WakeIOThread();
     UnlockChipStack();
 
     osStatus err = osOK;
