+++ src/system/SystemConfig.h
@@ -111,6 +111,14 @@
 
 // clang-format off
 
+#ifndef CHIP_SYSTEM_CONFIG_USE_IO_THREAD
+#if CHIP_SYSTEM_CONFIG_USE_SOCKETS || CHIP_SYSTEM_CONFIG_USE_NETWORK_FRAMEWORK
+#define CHIP_SYSTEM_CONFIG_USE_IO_THREAD 1
+#else
+#define CHIP_SYSTEM_CONFIG_USE_IO_THREAD 0
+#endif // CHIP_SYSTEM_CONFIG_USE_SOCKETS || CHIP_SYSTEM_CONFIG_USE_NETWORK_FRAMEWORK
+#endif // CHIP_SYSTEM_CONFIG_USE_IO_THREAD
+
 /**
  *  @def CHIP_SYSTEM_CONFIG_TRANSFER_INETLAYER_PROJECT_CONFIGURATION
  *
