+++ src/system/tests/TestSystemWakeEvent.cpp
@@ -33,7 +33,7 @@
 #include <support/UnitTestRegistration.h>
 #include <system/SystemError.h>
 #include <system/SystemLayer.h>
-#include <system/WatchableSocket.h>
+#include <system/SystemSockets.h>
 
 #if CHIP_SYSTEM_CONFIG_POSIX_LOCKING
 #include <pthread.h>
@@ -42,17 +42,6 @@
 using namespace chip::System;
 
 #if CHIP_SYSTEM_CONFIG_USE_SOCKETS
-
-namespace chip {
-namespace System {
-class WakeEventTest
-{
-public:
-    static int GetReadFD(const WakeEvent & wakeEvent) { return wakeEvent.GetReadFD(); }
-};
-} // namespace System
-} // namespace chip
-
 namespace {
 
 struct TestContext
@@ -76,15 +65,15 @@ struct TestContext
         FD_ZERO(&mReadSet);
         FD_ZERO(&mWriteSet);
         FD_ZERO(&mErrorSet);
-        FD_SET(WakeEventTest::GetReadFD(mWakeEvent), &mReadSet);
-        return select(WakeEventTest::GetReadFD(mWakeEvent) + 1, &mReadSet, &mWriteSet, &mErrorSet, &timeout);
+        FD_SET(mWakeEvent.GetNotifFD(), &mReadSet);
+        return select(mWakeEvent.GetNotifFD() + 1, &mReadSet, &mWriteSet, &mErrorSet, &timeout);
     }
 };
 
 void TestOpen(nlTestSuite * inSuite, void * aContext)
 {
     TestContext & lContext = *static_cast<TestContext *>(aContext);
-    NL_TEST_ASSERT(inSuite, WakeEventTest::GetReadFD(lContext.mWakeEvent) >= 0);
+    NL_TEST_ASSERT(inSuite, lContext.mWakeEvent.GetNotifFD() >= 0);
     NL_TEST_ASSERT(inSuite, lContext.SelectWakeEvent() == 0);
 }
 
@@ -96,11 +85,11 @@ void TestNotify(nlTestSuite * inSuite, void * aContext)
     // Check that select() succeeds after Notify() has been called
     lContext.mWakeEvent.Notify();
     NL_TEST_ASSERT(inSuite, lContext.SelectWakeEvent() == 1);
-    NL_TEST_ASSERT(inSuite, FD_ISSET(WakeEventTest::GetReadFD(lContext.mWakeEvent), &lContext.mReadSet));
+    NL_TEST_ASSERT(inSuite, FD_ISSET(lContext.mWakeEvent.GetNotifFD(), &lContext.mReadSet));
 
     // ...and state of the event is not cleared automatically
     NL_TEST_ASSERT(inSuite, lContext.SelectWakeEvent() == 1);
-    NL_TEST_ASSERT(inSuite, FD_ISSET(WakeEventTest::GetReadFD(lContext.mWakeEvent), &lContext.mReadSet));
+    NL_TEST_ASSERT(inSuite, FD_ISSET(lContext.mWakeEvent.GetNotifFD(), &lContext.mReadSet));
 }
 
 void TestConfirm(nlTestSuite * inSuite, void * aContext)
@@ -110,7 +99,7 @@ void TestConfirm(nlTestSuite * inSuite, void * aContext)
     // Check that select() succeeds after Notify() has been called
     lContext.mWakeEvent.Notify();
     NL_TEST_ASSERT(inSuite, lContext.SelectWakeEvent() == 1);
-    NL_TEST_ASSERT(inSuite, FD_ISSET(WakeEventTest::GetReadFD(lContext.mWakeEvent), &lContext.mReadSet));
+    NL_TEST_ASSERT(inSuite, FD_ISSET(lContext.mWakeEvent.GetNotifFD(), &lContext.mReadSet));
 
     // Check that Confirm() clears state of the event
     lContext.mWakeEvent.Confirm();
@@ -147,7 +136,7 @@ void TestClose(nlTestSuite * inSuite, void * aContext)
     TestContext & lContext = *static_cast<TestContext *>(aContext);
     lContext.mWakeEvent.Close();
 
-    const auto notifFD = WakeEventTest::GetReadFD(lContext.mWakeEvent);
+    const auto notifFD = lContext.mWakeEvent.GetNotifFD();
 
     // Check that Close() has cleaned up itself and reopen is possible
     NL_TEST_ASSERT(inSuite, lContext.mWakeEvent.Open(lContext.mWatchableEvents) == CHIP_NO_ERROR);
